feat(leave): M338 — leave category master + balance ledger (Leave PRD §3/§8)
V185: leave_mgmt.leave_category (7 seeded: PAID/UNPAID/MEDICAL/STATUTORY/FAMILY/EDUCATIONAL/BUSINESS)
       + leave_mgmt.leave_balance_ledger (append-only, 11 tx types, balanceAfter snapshot)
       + category_id FK on leave_type.

Single-write-point LeaveLedgerService.record() wires into all balance mutations:
LeaveBalanceService (OPENING/RESERVATION/LEAVE_TAKEN/LEAVE_CANCELLED/ADJUSTMENT),
LeaveAccrualService (ACCRUAL), LeaveYearEndService (CARRY_FORWARD/EXPIRY).

Also fixes pre-existing schema mismatch: AttendancePayrollSummary.period_year/month
declared short (SMALLINT) to match V184 DDL.

SPA: LeaveCategoriesPage (inline CRUD), LeaveBalancesPage gains "Ledger" drawer,
"leave/categories" nav entry (TagsOutlined), en/az i18n.

Boot-verified V185 applied; smoke: GET /api/leave/categories → 7 rows ✓,
GET /api/leave/balances/ledger?employeeId=...&year=2026 → [] ✓

+++ src/main/java/az/millers/hcm/attendance/domain/AttendancePayrollSummary.java
@@ -39,10 +39,10 @@ public class AttendancePayrollSummary {
     private UUID employeeId;
 
     @Column(name = "period_year", nullable = false)
-    private int periodYear;
+    private short periodYear;
 
     @Column(name = "period_month", nullable = false)
-    private int periodMonth;
+    private short periodMonth;
 
     @Column(name = "total_late_minutes", nullable = false)
     private int totalLateMinutes;
